Eliminar e insertar funcionales

+++ Interface/Interface/Eliminar.xaml.cs
@@ -28,7 +28,8 @@ namespace Interface
 
             client = new puntoVentaSoapClient();
 
-            productos.ItemsSource = client.readProductos();
+            actualizar();
+           
         }
         private void Border_MouseLeftButtonDown_1(object sender, MouseButtonEventArgs e)
         {
@@ -44,14 +45,31 @@ namespace Interface
             frm.Show();
         }
 
+        void actualizar()
+        {
+            var linq = from elemento in client.readProductos()
+                       select new
+                       {
+                           Nombre = elemento.nombre,
+                           Precio = elemento.precio,
+                           Stock = elemento.stock,
+                           Descripción = elemento.descripcion
+                       };
+            productos.ItemsSource = linq.ToList();
+        }
+
         private void button_Click(object sender, RoutedEventArgs e)
         {
             try
             {
                 Productos eliminado = productos.SelectedItem as Productos;
 
+                if (eliminado == null) return;
+
                 client.deleteProductos(eliminado.id);
                 MessageBox.Show("Eliminado.");
+                actualizar();
+                
             }
            catch(Exception ex)
             {
